Add: sync Estate table to queue
