+++ package.json
@@ -25,6 +25,7 @@
     "@apollo/server": "^4.10.0",
     "@nestjs/apollo": "^12.0.11",
     "@nestjs/axios": "^3.0.1",
+    "@nestjs/bull": "^10.1.0",
     "@nestjs/cache-manager": "^2.2.0",
     "@nestjs/common": "^10.0.0",
     "@nestjs/config": "^3.1.1",
@@ -37,6 +38,7 @@
     "@nestjs/typeorm": "^10.0.1",
     "axios": "^1.6.5",
     "bcrypt": "^5.1.1",
+    "bull": "^4.12.2",
     "cache-manager": "^5.4.0",
     "cache-manager-redis-store": "^2.0.0",
     "class-transformer": "^0.5.1",
